refactor(user): 관심사 선택 API의 Profile 도메인 분리
UserController와 UserService에 혼재되어 있던 관심사 선택 관련 API를 Profile 도메인으로 이전하여 책임과 역할 분리

- UserController -> ProfileController로 API 이동
- UserService -> ProfileService로 관련 비즈니스 로직 이동
- 유저 관심사를 저장하기 위한 UserInterest 엔티티 추가
