+++ src/main/java/piq/piqproject/domain/users/controller/ProfileController.java
@@ -1,14 +1,20 @@
 package piq.piqproject.domain.users.controller;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import piq.piqproject.common.list.ListResponseDto;
+import piq.piqproject.domain.users.dto.request.UserInterestRequestDto;
+import piq.piqproject.domain.users.dto.response.UserInterestResponseDto;
 import piq.piqproject.domain.users.entity.UserEntity;
 import piq.piqproject.domain.users.service.ProfileService;
 
@@ -54,5 +60,20 @@ public class ProfileController {
         profileService.deleteVoice(userEntity.getId());
 
         return ResponseEntity.ok("Voice deleted successfully.");
+    }
+
+    /**
+     * 현재 로그인된 사용자의 관심사를 저장하는 API입니다.
+     * 
+     * @param userEntity             @AuthenticationPrincipal을 통해 주입된 현재 인증된 사용자 엔티티
+     * @param UserInterestRequestDto 유저가 선택한 관심사 리스트를 담은 request dto
+     * @return 저장된 사용자의 관심사 리스트
+     */
+    @PreAuthorize("hasRole('ROLE_USER')")
+    @PostMapping("/me/interests")
+    public ResponseEntity<ListResponseDto<UserInterestResponseDto>> registerUserInterests(
+            @AuthenticationPrincipal UserEntity user,
+            @Valid @RequestBody UserInterestRequestDto userInterestRequestDto) {
+        return ResponseEntity.ok(profileService.registerUserInterests(user, userInterestRequestDto));
     }
 }
